LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-meta-refresh-expected.txt



FAIL <meta> refresh fires navigate event assert_equals: expected "reload" but got "push"
